Move parcelables in "models"
