+++ app/src/main/java/com/example/android/spotifystreamer/models/ParcelableTrack.java
@@ -1,4 +1,4 @@
-package com.example.android.spotifystreamer;
+package com.example.android.spotifystreamer.models;
 
 import android.os.Parcel;
 import android.os.Parcelable;
